If a house ID is not included in the request to retrieve characters or no characters are returned from the database, the application now displays a message on the characters page instead of returning an error code that makes an error appear. The models have been separated into distinct files
