+++ GameOfThrones/GameOfThrones.csproj
@@ -142,6 +142,8 @@
     </Compile>
     <Compile Include="Migrations\Configuration.cs" />
     <Compile Include="Models\Castle.cs" />
+    <Compile Include="Models\Character.cs" />
+    <Compile Include="Models\House.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
